Refactor into distinct package for licensing purposes.
