+++ pom.xml
@@ -1,5 +1,6 @@
 <!--
  ~ Copyright 2019 Gluon
+ ~ Copyright 2020 NMRbox.org
  ~
  ~ Licensed under the Apache License, Version 2.0 (the "License");
  ~ you may not use this file except in compliance with the License.
@@ -12,20 +13,23 @@
  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
-  -->
+ ~ 
+ ~ NMRbox.org extension funded by the National Institutes of Health (NIH) /
+ ~ National Institute of General Medical Sciences (NIGMS), grant 1P41GM111135 
+ -->
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.openjfx</groupId>
-  <artifactId>javafx-maven-plugin</artifactId>
+  <groupId>edu.uchc.connjur</groupId>
+  <artifactId>connjurpackage-maven-plugin</artifactId>
   <packaging>maven-plugin</packaging>
-  <version>0.0.6</version>
-  <name>javafx-maven-plugin Maven Mojo</name>
-  <description>The JavaFX Plugin is used to run JavaFX 11+ projects</description>
-  <inceptionYear>2019</inceptionYear>
+  <version>1.0.0</version>
+  <name>connjurpackage-maven-plugin Maven Mojo</name>
+  <description>Plugin used to package JavaFX 11+ projects</description>
+  <inceptionYear>2020</inceptionYear>
 
-  <url>https://openjfx.io</url>
+  <url>https://www.nmrbox.org</url>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
